+++ R/export.R
@@ -769,7 +769,7 @@ as_word_tbl_body <- function(
 #' 13-10
 #'
 #' @section Function Introduced:
-#' `v1.0.0` (March 26, 2026)
+#' `v1.4.0`
 #'
 #' @export
 as_typst <- function(